requirements.txt
streamlit==1.35.0
openai==1.30.3
newspaper3k==0.2.8
